requirements.txt
flask==3.1.3
google-genai==1.64.0
Pillow==11.0.0
certifi==2024.12.14
gunicorn==23.0.0
